+++ lib/utils/rafts.ts
@@ -1,0 +1,8 @@
+export const boatOptions = [
+    { value: 'single-kayak', label: 'Single Kayak' },
+    { value: 'double-kayak', label: 'Double Kayak' },
+    { value: 'small-raft', label: 'Small Raft' },
+    { value: 'round-raft', label: 'Round Raft' },
+    { value: 'medium-raft', label: 'Medium Raft' },
+    { value: 'large-raft', label: 'Large Raft' }
+];
